requirements.txt
alembic
geopy
pandas
pycountry
python-dateutil
python-dotenv
requests
sqlalchemy
